New dbc Class and updates to existing classes
Added dbc class to expedite connection to the database. NOTE: this
requires the config.ini file for the site to be stored in the "secure"
directory on the same level as the "www" directory. i.e. the config.ini
file will NOT be in the Git repository, as it contains database
connection information (which we don't want on the public repo).