Permission, Controller, Authenticator, and Dbc work
Added new Permissoin class to help in authenication and page-related
permission control. Added file for the Logger class so we have it.
Worked on implementing registration (createAccount) for the
Authenticator, and now it should just be up to the Controller to finish
implementation for that use case. Also updated auto CSS and JavaScript
HTML generation functions to be easier to use. Further updated User
class to handle user permissions.