Finished UC2 implementation
While it was a bit of a pain to debug this as I was going through it
(partly my fault for not using exceptions properly/liberally enough),
I've finished implementation of UC2. Currently, there is no visual
indication of success on the UI, but you might get some horrible-looking
error message when it fails. You can see the "user" variable has been
set in the $_SESSION though if you var_dump it out. No form validation
in JS on the page, but some validation does take place in the PHP
classes.